+++ spec/lib/renogen/formatters/base_spec.rb
@@ -1,25 +1,25 @@
 require 'spec_helper'
 
 describe Renogen::Formatters::Base do
+  let(:version) { '1.2.34' }
+  let(:date) { Date.today }
+  let(:changelog) { double(Renogen::ChangeLog::Model, version: version, date: date) }
+
   describe '#header' do
     subject { described_class.new.header(changelog) }
-
-    let(:version) { '1.2.34' }
-    let(:date) { Date.today }
-    let(:changelog) { double(Renogen::ChangeLog::Model, version: version, date: date) }
 
     it { is_expected.to eql "#{version} (#{date})" }
   end
 
   describe '#write_header' do
-    it 'raises an NotImplementedError' do
-      expect{subject.write_header('header')}.to raise_error NotImplementedError
+    it 'does nothing' do
+      expect(subject.write_header(changelog)).to be_nil
     end
   end
 
   describe '#write_group' do
-    it 'raises an NotImplementedError' do
-      expect{subject.write_group('group_name')}.to raise_error NotImplementedError
+    it 'does nothing' do
+      expect(subject.write_group(changelog)).to be_nil
     end
   end